ai-engine/requirements.txt
# ====================================================
# AI Engine: OCR / Document Intelligence Dependencies
# ====================================================

# --- PDF Text Extraction ---
PyMuPDF           # fitz: fast native text extraction from digital PDFs (REQUIRED)
pdfplumber        # alternative PDF extraction with table support (OPTIONAL)

# --- Scanned PDF → Image Conversion ---
pdf2image         # converts scanned PDFs to images for OCR (REQUIRED for handwriting)
# Windows: also requires Poppler binaries in PATH

# --- OCR Engine ---
easyocr           # deep-learning OCR supporting English + Hindi (REQUIRED)
# Note: Downloads ~200MB model on first run

# --- Image Processing ---
Pillow            # PIL image opening, format conversion (REQUIRED)
opencv-python     # image preprocessing: deskew, denoise, threshold (OPTIONAL but recommended)
numpy             # array ops for opencv/easyocr image pipeline (REQUIRED)

# --- Language Detection ---
langdetect        # lightweight ML language identifier (REQUIRED for detect_language.py)

# --- Environment ---
python-dotenv     # reads .env config files (REQUIRED)

# --- AI & NLP ---
google-generativeai   # Gemini API for translation and structuring
sentence-transformers # Semantic embedding for clustering
scikit-learn         # DBSCAN clustering
fastapi              # API framework
uvicorn              # Web server
requests             # Webhooks
pymongo              # MongoDB connection
